Allow a Contact to be marked as the incoming trust main contact
Now that we have extended the Create Contact FormObject to support
`incoming_trust_main_contact`, we can add a checkbox to the new & edit forms
to enable us to mark a contact as the "main contact" for an incoming trust.

We also show if a contact is the incoming trust's main contact on the contacts
listing page.