Add a basic search service
Search is hard, we accept that, but it is also extremely valuable to our
users.

We have a plan to deliver the search facility we know users value, but
we do not have the foundations yet. This work begins to lay them

Here we add a search service based on what we can search for:

- a URN
- a UKPRN

As we store these identifiers in our tables, we can quickly search them.

We don't intend to put this feature infront of users, we'll do that once
we can search by the school or academy name, whcih is coming.

But we do want to set the approach for our searching, which is here.

We've kept things a simple as we can whilst hinting that more is to
come.

We look at the query string and infer what kind of query is going to
deliver a result, loosely:

- a 6 digit number is likely a URN
- an 8 digit number is likely a UKPRN
- anything else we cannot help right now

We have two UKPRN values, so complexity already begins to rise, we
search both and return matches as this is the likely expectation of
users.

For now, this is all we need, we test the service and will come back
once we can add school and academy name search based on the new
Gias::Establishment model.