Allow a Contact to be marked as the outgoing trust main contact
Now that we have extended the Create Contact FormObject to support
`outgoing_trust_main_contact`, we can add a checkbox to the new & edit forms
to enable us to mark a contact as the "main contact" for an outgoing trust.

We also show if a contact is the outgoing trust's main contact on the contacts
listing page.