Change how GIAS::Establishment codes are stored
After testing the Gias::Establishment importer it became clear that most
of the 'code' values are padded with zeros e.g. 611 becomes 0611.

The database was doing the right thing and coercing the values into
integers, but the next import would update every record because the
values are compared as strings so "611" != "0611". The upshot is every
record was updated every import - which is not what we want.

Another thing that was noted is that the raw csv from GIAS contains the
padding, but opening and saving in an application like MS Excel silently
strips the padding - doing so results in the same issue, with the
comparison triggering an update.

There are lots of way to approach this, on balance, we felt that we
should just store the code as a string. Our reasoning is:

- the padding is present in the original source
- the data is for reference only
- we prioritise a clean and simple import
- we are surfacing these codes in the UI and nothing more (potentially
  not even showing codes to users)
- we should not take on the responsibility for the quality of the data
  and make attempts to fix it unless absolutely necessary

Therefore this work is a migration for integer to string values.