Update Gias::Establishment importer
During testing with the real dataset we soon learnt returning the
results of what records get update and with what values is really
useful.

There is one issue with this: a new record is also an update because we
cannot upsert - we don't think this matters too much as the results are
not meant for end users right now, but it is something to keep in mind.