Update Your team projects in progress table to display Team or Region
Regional case workers need to be able to see an overview of all the
regions projects and so this work is to ensure that if the current user
is a Regional case worker, they will see what region each project
belongs to. For all regional delivery officers, they will only see the
projects which are within their same team.