Add basic event model
We want to be able to offer users a kind of 'audit log' across the
application, here we lay the foundation for that by adding an `Event`
model to store the individual log entries.

For now we have a Project, as the main data of interested and then a
generic 'system' event for anything else.

We will likely add more assoications as we learn more about what we want
to keep an audit of, which this model supports via the `eventable`
assoication.

Any other ActiveRecord model can opt into event logs by declaring the
association.

Specific groupings can also be handled by the `grouping` enum to collect
logical event together, for example a log of the import process might be
grouped in `import`.

We are not yet triggering events - that will come later.