Allow contacts to be editbale or not
We have different classes of contacts, some can be edited by the user
manage a project to which they relate and some cannot.

Here we encapsulate this on each class.

This lets the view only render the 'edit' UI on editable contacts -
further authorisation based on user, etc can still be done as needed.

Along the way we cover the base Contact class with a spec and only test
the difference in the Contact::Project subclass.