Fix and update Contact::Project
The concept of a 'project contact' is a set of contact details that
relate to a project, the association to Project is the foundation of the
object so should be required.

Because the Contact subclass is called 'Project' we have to be careful
when dealing with the related Project class, by explicitly declaring the
assoication class_name as '::Project' we tell Rails it is not the
Contact::Project. We chose this over the slightly imporved renaming the
class to something like `Contact::ProjectContact` as this would require
updating the `type` column in the database.

With both these changes asking a Contact::Project if it is a particular
role for the project becomes much simpler.