Fix and update 'primary' contact behaviour
We want users to be able to indicate who the 'primary' contact for the
following categories is:

- school or academy
- incoming trust
- outgoing trust

The existing implementation was updating all three associations
everytime the form was submitted, this resulted in any existing primary
contacts being lost.

Whilst working to fix this, we felt like the whole concept could be
simplified by assuming the category and setting the primary contact,
with the result being a single 'is this the primary contact for the
choosen category' checkbox.

Whilst this is probably not the ideal experience for this, it does fix
the issue and sidesteps and data issues, leaving the path clear to come
back and work on the experience later.

We also refactor the form object, by passing in either a new empty
Contact or an Existing one, we can loose a lot of the complexity.

We switch from the word 'main' as it clashes with the 'project main
contact' which is something else, we could fix everything related to
that here as we felt it was too much, the associations for example, but
this strikes a balance that is good enough.

We fix the original issue by checking the category and setting only the
realated association on the Project.