Override User.add_new_project database column
We want to temporarily grant RCS users the ability to add new projects
(specifically Transfer projects, but all projects for convenience).

To do this, we have overridden the `add_new_project` column in the db with
a method. The db column is still set to `false` on creation of an RCS user,
but the overriding method returns true.