Move the site banner environment variable away from sentry
We used sentry for our application performance monitoring and erroring
insights, but we are now required to use App Insights. This means that
we will be dropping support for sentry and removing it from the service.
This work is to remove all dependencies on sentry within the environment
banner scope.