Use a form object to call the Import::GiasEstablishmentImportJob
This form object takes an uploaded file (an ActionDispatch::Http::UploadedFile)
and a user, and called Import::GiasEstablishmentImportJob with a time delay.

The time delay is set via an environment variable, but defaults to 0400 the
next day.

The form validates that the uploaded file is not empty, is not larger than 150Mb
, is of type text/csv and has the expected columns.

If the file is not valid, the file is not saved to the filesystem.