Allow service support users to upload a file of GIAS data for ingestion
Add a very basic (needs content design!) form to allow users to upload GIAS
data for ingestion later (via the Import::GiasEstablishmentImportJob)

If the uploaded file is valid, the user will see a message saying their file
will be uploaded at 4am the next morning and that they will receive an email.

If the file is not valid, the user will see a flash message with the validation
errors.

Note that if the ingestion fails later on, the user will receive the errors in
an email, they will not see the errors on this page.