Add Contact::Establishment model
This model is for Contacts who belong to an establishment. They have a
mandatory `establishment_urn` field.

This has not been created using standard Rails associations (belongs_to/has_one)
to make it easier to fetch the contacts associated with a Project. A project has
an establishment via a URN, butthis establishment is *not* a Gias::Establishment
model, it is instead fetched from an API.

So if we were to associate the new
Contact::Establishment with a Gias::Establishment via a foreign_key association,
when it came to fetching the contacts for a project, we would not be able to use
`project.establishment.contact` because the `establishment` associated with the
project is not actually a Gias::Establishment.

This works for now but we can add "true" rails associations later if the
establishment association with projects becomes a Gias::Establishment.