Fix: Update GIAS job handling
We change three main things here:

We recently lowered the resolution of the file name timestamp as we were
seeing them not matching up with the two calls, but even with seconds,
it was still possible to make the second `file_path` call in the 'next'
second, making the job be queued with the wrong filename.

To remove this issue all together, we assign the file name and use that
instead, this guarentees that the filename is the same.

We fix a bug in the results email which added one too many 'results'
causing nil errors.

We also change the keys of the results hash to be symbols instead of
integers as ActiveJob does not support integer as hash keys:

https://guides.rubyonrails.org/active_job_basics.html#supported-types-for-arguments