Map contact attributes from the CSV rows
We need to extract the Head Teacher contact data from the CSV rows, and map
them in a way so that we can create new Contact::Establishment records from
them.

We add the `category` attribute manually, as it will always be `school_or_academy`

We also needed to do some magic with the Head teacher's name, as the CSV divides
first and second name up, but the contact expects a single name.