Extract and update Headteacher importer
Previously we attempted to import the headteacher contact at the same
time as the establishments themselves. This was a good approach, but
when we tested it in production like environments, the memory usage was
high and was possibly causing the container to fail and restart.

We decide to change the approach, as we have a long time window to import
data, we decided to extract the headteacher import from the
establishments and run them seperately.

We took the opportunity to update the service at the same time,
swtiching from storing the errors in the results hash to logging them -
this should give us better visibility of the import process.

We introduced the idea of required values so that we can skip the row if
it does not contain the values we must have to create a new contact.

If the source is missing the 'title' or 'HeadPreferredJobTitle' we
provide a default of 'Headteacher'.

We have not removed the old code or created a ActiveJob yet, that comes
next. The `Import::GiasEstablishmentCsvImporterService` spec needed to
be tweaked as we added a row to the fixture file.