Model Contact::Parliament
Model the Member of Parliament contact like the other contact types. However,
this contact doesn't persist in our database - they are fetched from the
Members API via the constituency name, and only modeled as a transient object.

If there are any errors from the Members API, raise them. We will have to ensure
we handle these errors in the relevant controllers.