(Fix) Raise NotFoundError when search term returns no results
When working on `Contact::Parliament`, we found that a search term which returns
no resolts doesn't raise an appropriate error in the MembersApi client. Instead
of a NotFoundError, searching for an unknown consttiuency search term was raising
a NomethodError on `constituency_data.object["items"]`.

This raises a NotFoundError instead, so we can catch it further downstream.