Add job for headteacher import
After the GIAS establishment import has run, the same file can be used
to update or create any headteacher details it contains.

Here we add and queue a new job one hour after the establishment job
runs.

Becuase we want to use the same file, we only delete it once the second
job has run - whilst there could be a race condition of sorts here,
both import services handle the file not being found gracefully and we
are not concerned about any condition at this point.

We will add a second email for the headteacher contacts next.