Add a GIAS headteacher import email
We want the user who initiated the import to know that it complted and
some results of the import job, just like the establishment import.

Here we add this bechaviour along with a basic GOV.UK Notify template to
allow the application to send the email.

Along the way we make some tweaks to the existing implementation.