Change the GIAS Import time
The application now runs with it's own database instance so there is no
need to try and wait for a 'quiet' time to run out import jobs.

They still take a significant amount of time, so running them in the
background is still required.

Here we bring the time forward to 0100.