Do not include the changes hash in the email job
After importing GIAS establishment and headteacher contacts we send an
email to the user that summerises the import.

When queuing the email we pass the whole `result` object from the import
service - this has the potential to be very large e.g. on inital import
of 50K rows, you can expect 50K x number of imported columns.

On testing in the development environment, we suspect this hash is
unreasonably large for Sidekiq to serialise.

As we do not include the changes in the email anyway, here we remove the
changes before queuing the job.