Search by words or word-like strings
We want to be able to search for a project by establishment name. To do this,
we first have to identify that the query string is a "word like" string and
not a URN or UKPRN.

Once that is done, we search the GIAS::Establishment records using the query
against the Establisment names. If we find any matches, we use the URNs of the
matching establishments and pass them to `search_by_urns`.