Add Gias::Group model
A Gias Group is what we currently think of as a "trust", but the data from GIAS
contains more than just trusts, so for safety's sake we are calling them Groups.

Note that the Companies House number is being stored as a string, because it can
contain leading zeros which are significant. If we stored it as an integer, Rails
would "helpfully" strip off leading zeroes.