Add 2RI to new transfer form
We already have this question about 2RI on a conversion and we have
learnt that it also applies to transfers.

Here we add the question to the form and expose the response on the
'About the project' tab on a project.

This is all relatively straight forward thanks the the attirubue being
on our Project class and so shared across both types.