Add a basic view for the Grant management and finance unit export page
This page shows a list of the previous 12 months and allows the user to download
a list of projects whose Advisory board dates are in those months.

The other export pages show the next 6 month,s but because Advisory board dates
are in the past, we should show the current month & previous 11 (to make up a
full year)

This page is currently ont linked from anywhere and is not accessible to any
particular user group yet.