Build the export journey to download all Transfers by month
Now that we have the Transfers by month CSV export, build a journey for the
user to access this export.

First, the user goes to an index page where they can see a list of the current
and next 6 months. From there, they can click through to the existing "by month"
page views to see all confirmed and revised projects for those months.

Then, the user can click through to an export page, and click a button to
download all transfers (confirmed and revised) for that month.