Remove width constraint from paragraphs
In commit https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/commit/7d1c2fe5ae3188545d045dc8d9f87a1fc5c9aa54#diff-8fe1b0ed3156a6d8f6026964c5283e72ad5bf95cec2c2609199c0e1d5867024bR5 a width constraint was added to many elements in the
application. The reasoning behind this was to keep all content to a readable
length in the new 1200px wrapper. The dimensions for the readable length were
taken from https://baymard.com/blog/line-length-readability

However, this has had an effect whereby some paragraph content has become too
narrow and looks too "heavy" on the page.

We are removing this width constraint for now, however we are aware that if
there is an assessment we may have to re-apply it if the width is not deemed
optimal for readability.