Add Date received field to the Conversion Receive grant payment certificate task
We need to be able to enter the date the Grant payment certificate was received
to the Conversion task list. This was not a simple piece of work as we do not
yet have any tasks that collect dates.

Refactor the Grant payment certificate task form to collect a date, or show an
existing date if one was previously entered. We have also modified the
Project#grant_payment_certificate_received? method to take the date received
into account.