Split the `check_and_save` in Conversions Receive grant payment certificate task
We want to split the existing `check_and_save` action into `check` and `save`.

To do this, we have repurposed the existing `check_and_save` action into `save`
(actually `save_certificate` due to save being a reserved word). We have also
added a data migration to update the value of the new `check` action to be the
same as `save` - we can comfortably assume that if someone marked the
`check_and_save` action as true before, then both `check` and `save` are true.

We then add the new checkboxes and text to the edit form page.