Script to backfill Grant payment certificate dates for live data
In https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/1283
we added a new date field to the Conversion Grant payment certificate received
task.

Some Conversion projects are already completed in production and therefore
cannot be edited by users. We need a way to backfill these projects with their
relevant dates.

A spreadsheet has been supplied with URNs of the academies, and the date the
grant payment certificate was received. Once this script is deployed, we can run
the associated rake task with this CSV (uploaded to Azure) and update the
relevant projects.