Enable service support users to be able to update all projects
The service support user does a lot of data quality work in KIM and
updates projects within this service on behalf of regional delivery
officers when there is an error. This change will allow this process to
happen without needing the assigned user to be present and make the
process a lot simplier.