Remove notification banner for service support users
When a user views a project that they are not assigned to, a
`Not assigned to project` notification banner is displayed in order to
communicate to the user that they cannot edit this project. As this now
doesn't apply to service support users, who can edit any project, this
notification banner has been removed.