Add `is_academy?` to Gias::Establishment & Api::AcademiesApi::Establishment
We need to know if an establishment (of either API type or local model type) is
an academy. use a best-guess list of codes from the Academies db to see if an
establishment is likely to be an academy.

These codes are:

46	Academy 16 to 19 sponsor led
45	Academy 16-19 converter
42	Academy alternative provision converter
43	Academy alternative provision sponsor led
34	Academy converter
44	Academy special converter
33	Academy special sponsor led
28	Academy sponsor led
57	Academy secure 16 to 19

I'm not happy with the amount of duplication here but we can refactor later.