Namespace the export headers by Project type
We need to change one of the column headers on exports for Transfer projects.

Before, all column headers were the same for both conversions and transfers. Now
we need the header for "all conditions met" to change between transfers and
conversions.

To do this, first we need to namespace the headers, so later on we can change
one for transfers only.

Duplicate all the headers and namespace them under `transfers` and `conversions`
Add a variable to the export services to indicate which project type we're
generating a report for.
Then, add the project type to the i18n strings to fetch the correct string for
each project type.