Change "All conditions met" header in exports for Transfer projects
The header for the "All conditions met" column in the Transfers export should be
"Authority to proceed"

The simplest way to do this is to add a `authority_to_proceed` alias to the
Project Presenter, and use this to return the `all_conditions_met` value
for the Project. Because the name of the method is `authority_to_proceed`, we
can add a new I18n string to the file with the correct header.

Therefore when we call the column header `authority_to_proceed` in the export,
we get the I18n string for `authority_to_proceed` as the header, and the value
of `all_conditions_met?` in the value.