Add scope to projects to filter by significant date in a range of dates
We already have scopes to return projects by their significant date. We now
need to return projects with a significant date in a range of dates, i.e.
date from - date to.

To do this we have added a new class method (which acts like a scope in a
concern) to get all projects with confirmed dates, and filter them by their
sognificant date being within the desired range.