Bugfix: Return the correct main contact details in the project presenter
we had a bug report that the exports did not always contain the correct main
contact details for the school. While the `Project.main_contact` association
works fine in the UI the project presenter pattern is always returning the
first Contact association, and not the one indicated by the
`main_contact_id` attribute.

To temporarily fix this, return the main contact details in the project
presenter a different way. We will have to circle back to this and figure out
why the associations aren't working as expected at a later date.