New By month view for non-data-consumer users (non export users) - Conversions
Add a new view for non-data consumer users to view a list of Conversion projects
with a significant date in the upcoming month. The table view is identical to
the date range view added in https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/1330

Non-data consumer users do not see the date range picker, and instead only see
the upcoming month. We have built in the ability for a different month to be
viewed by manipulating the URL.