By month view for Caseworkers & other non-data consumer users - Transfers
Add a new view for non-data consumer users to view a list of Transfer projects
with a significant date in the upcoming month. The table view is identical to
the date range view added in #1338

Non-data consumer users do not see the date range picker, and instead only see
the upcoming month. We have built in the ability for a different month to be
viewed by manipulating the URL.