Add new trust name & reference number to projects (for Form a MAT)
We need to add a new project type of "Form a MAT". Due to time constraints, we
are adding this in as stripped-back a form as possible. Form a MAT projects are
Conversion or Transfer projects which are grouped together under a new Trust.
As this Trust does not yet exist in GIAS, we need to store its name and TRN
somewhere - so we will store it for now on the projects.

As the Trust doesn't exist in GIAS, we cannot reference it using the UKPRN from
GIAS/Academies API.