(Fix) Return all other contacts for a project in CSV via ContactsFetcherService
In https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/1366
we used the ContactsFetcherService to correctly return all contacts for a project
in order to pull out the local authority and school/academy contacts. We now
need to duplicate this work for Trust contacts, and for completeness we have
used the same method to get the Diocese and Solicitor contacts this way.

This means that if a contact is not a Contact::Project but instead a
Contact::Establishment contact, we still get them and put them in the CSV.