Imporve Sonarcloud config
A lot of the 'bugs' (11 of 19) reported by Sonarcloud are in the DfE frontend,
which we consider vendor code i.e. we don't update it. It should really
be consumed from the npm package, but this was broken when we added it.

Here we are exluding the DfE Frontend code from the Sonarcloud scanner.