Add bind mount for coverage
When CI runs all the checks it uses this docker-compose file, here we
add a bind mount which is a one to one mapping of a directory inside the
running container to the host it is running on.

The upshot is that when Rspec and Simplecov run inside the container,
the output goes to the host, where we can reach it.