Update Sonarcloud config
Now we can get at the coverage report we need to make sure Sonarcloud
can 'see' it correctly - not an easy task!

We download the Github artifact and copy it to the right location.

Because Sonarcloud is running inside a container, it took a lot of trial
and error to get the paths right, but we think we have it now.

Inside the Sonarcloud conatiner the work path is `/github/workspace/`,
we observed this is the logs in the Github action.

Knowing this, we `sed` the coverage report replacing the appropriat path
components from the container that ran the specs to that of the
Sonarcloud container.

We also need to tell Sonarcloud some things explicitally:

- to use the newer `coverage.json` file not the old `.ruleset` one
- to use the current directory as the work path
- that the specs (tests) are in the `spec` directory

With all that done, the coverage report should be in Sonarcloud!