Do not fetch trust if the Incoming trust UKPRN is nil
 We need to guard against instances where we try to get the Incoming trust
details for a project, but the incoming_trust_ukprn on a project is nil (i.e.
the project is a Form a MAT project).

When the incoming_trust_ukprn is nil, return a new Api::AcademiesApi::Trust
object, using the `new_trust_name` and `new_trust_reference_numver` to
instantiate the object. This means the pages which use this `incoming_trust`
object will still work as expected.