Fix unassigned project can render to csv
The assigned_to_name and assigned_to_email assumed that projects are
always assigend to a user, which is not technically the case.

We discovered this whilst testing one of the csv exports where we have
added a new column of data.

Adding a guard resolves this, outputting 'unassigned' as a more useful
indictator than a empty cell.