Add the date the academy opened conversion task
We want to capture the date the academy actually opened.

We lean on the existing receive grant payment certificate task to
validate the three parameter date, our use case is slightly different as
the date can be updated once set.

We also added some range validation on the year to help catch errors.

We did look to extract some of the validation to share among the form
objects, but on balance decided the duplication was better than the
miss-direction.

Co-authored-by: Lorna <lorna@dxw.com>