Temporarily remove the future date validation on project creation
We usually validate the provisional conversion/transfer date on a project to
always be in the future. However, as we are asking users to input historical
projects into Complete ahead of KIM shutting down, we need to loosen this
validation rule to allow projects with conversion/transfer dates in the past to
be added to the system.

We can re-introduce this validation at a later date.