Refactor ContactsFetcherService for performance
We do a lot of fetching contacts and this service is pretty happy to hit
the database a lot. In production with ~120 projects, exports are taking
a full minute to run and so the app is timing out.

We want to start by optimising our queries and this is the first
candidate (with more to follow).

The principles here are:

- there can generally be one ContactsFetcherService for one Project as
  it references the project internally
- our exports presenters are actually modules that all get mixed in to
  `project_presenter` so we make one instance of ContactsFetcherService
  available to all rather than instantiating (and so new db calls) new
  ones

This change has a big impact on the presenter specs as we have to stub
extra methods or (as we chose) switch to factories.

The result of this is, for a single conversion:

- 39 database queries per project/row in export before change
- 23 database queries per project/row in export after change

We'll have to see if this makes enough of a difference and we have
identified more optimisations we can make.