Cache the LocalAuthority instance for export
The performance of our exports is pretty poor, this work aims to further
reduce the number of database queries needed to generate the exports.

Here we load the Projects LocalAuthority when instantiating the
project_presenter, the mixed in LocalAuthorityPresenterModule then has
access to this single object to fetch any attributes rather than calling
the database each time.